The Meta Game
This is a little game about fashionable chameleons. There are three colors of chameleons. Red, Green and Blue. They all have their taste about what color they think is cool. a Red chameleon thinks Blue is cooler than Red so when it meets a blue chameleon it will change to Blue. a Blue will change to Green, and a Green will change to a Red. The graph on the bottom will show the color demographics of the population in real time. Eventually one color will win out.